requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
imagehash>=4.3.1
tqdm>=4.65.0
scikit-image>=0.21.0
